cleaned up the Climber subsystem and Climb command
