--- src/main/java/frc/robot/commands/Teleop/ClimbR.java
@@ -1,54 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands.Teleop;
-
-import edu.wpi.first.wpilibj.Joystick;
-import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.ClimberL;
-import frc.robot.subsystems.ClimberR;
-
-public class ClimbR extends Command {
-  ClimberR climber;
-  double flip;
-  
-  
-  /** Creates a new Climb. */
-  public ClimbR(ClimberR climber,double flip ) {
-    this.climber = climber;
-    this.flip = flip;
-    
-    addRequirements(climber);
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {}
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-    
-    
-
-
-      climber.setRightPower(flip);
-
-
-
-  }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-    
-    climber.setRightPower(0.0);
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
