+++ src/main/java/frc/robot/subsystems/Climber.java
@@ -11,12 +11,11 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.Robot;
 
-public class ClimberL extends SubsystemBase {
-  private TalonFX winchLeft; //winchRight;
+public class Climber extends SubsystemBase {
+  private TalonFX winchLeft, winchRight;
   
-
   //Enum for climber states
-  public enum climberPistonPosition{
+  public enum climberPistonPosition {
     FORWARDS, VERTICAL
   }
 
@@ -24,56 +23,49 @@ public class ClimberL extends SubsystemBase {
    * creates a climber object to control the climber on the robot
    * @author da bois
    */
-  public ClimberL() {
+  public Climber() {
   
     winchLeft = new TalonFX(Constants.ClimberConstants.LEFT);
-    //winchRight = new TalonFX(Constants.ClimberConstants.RIGHT);
+    winchRight = new TalonFX(Constants.ClimberConstants.RIGHT);
 
     winchLeft.setNeutralMode(NeutralMode.Brake);
-    //winchRight.setNeutralMode(NeutralMode.Brake);
-    
-    // winchLeft.setInverted(false);
-    // winchLeft.setSensorPhase(false);
-
-    // winchRight.setInverted(true);
-    // winchRight.setSensorPhase(true);
+    winchRight.setNeutralMode(NeutralMode.Brake);
 
     setPidsLeft(new double[] {0.4, 0, 0, 0});
-    //setPidsRight(new double[] {0.4, 0, 0, 0});
+    setPidsRight(new double[] {0.4, 0, 0, 0});
 
     winchLeft.configPeakOutputForward(0.75); 
-    //winchRight.configPeakOutputForward(0.75);
+    winchRight.configPeakOutputForward(0.75);
     winchLeft.configPeakOutputReverse(-0.75);
-    //winchRight.configPeakOutputReverse(-0.75);
+    winchRight.configPeakOutputReverse(-0.75);
 
-    
-		winchLeft.configMotionCruiseVelocity(9000);
-		winchLeft.configMotionAcceleration(3000);
+	winchLeft.configMotionCruiseVelocity(9000);
+	winchLeft.configMotionAcceleration(3000);
     winchLeft.configMotionSCurveStrength(0);
 
-		//winchRight.configMotionCruiseVelocity(9000);
-		//winchRight.configMotionAcceleration(3000);
-    //winchRight.configMotionSCurveStrength(0);
+	winchRight.configMotionCruiseVelocity(9000);
+	winchRight.configMotionAcceleration(3000);
+    winchRight.configMotionSCurveStrength(0);
 
     winchLeft.setInverted(true);
   }
 
   @Override
   public void periodic() {
-    //SmartDashboard.putNumber("Climber/Right encoder", getRightDistance());
+    SmartDashboard.putNumber("Climber/Right encoder", getRightDistance());
     SmartDashboard.putNumber("Climber/Left encoder", getLeftDistance());
-    //SmartDashboard.putNumber("Climber/Right Current", getRightCurrent());
+    SmartDashboard.putNumber("Climber/Right Current", getRightCurrent());
     SmartDashboard.putNumber("Climber/Left Current", getLeftCurrent());
   }
 
   public void cruiseVelocity(int encoderPerSecond) {
     winchLeft.configMotionCruiseVelocity(encoderPerSecond / 10);
-    //winchRight.configMotionCruiseVelocity(encoderPerSecond / 10);
+    winchRight.configMotionCruiseVelocity(encoderPerSecond / 10);
   }
 
   public void acceleration(int encoderPerSecondSqr) {
     winchLeft.configMotionAcceleration(encoderPerSecondSqr / 10);
-    //winchRight.configMotionAcceleration(encoderPerSecondSqr / 10);
+    winchRight.configMotionAcceleration(encoderPerSecondSqr / 10);
   }
 
   public void setPidsLeft(double[] pidf) {
@@ -83,17 +75,17 @@ public class ClimberL extends SubsystemBase {
     winchLeft.config_kF(0, pidf[3]);
   }
 
-  // public void setPidsRight(double[] pidf) {
-  //   winchRight.config_kP(0, pidf[0]);
-  //   winchRight.config_kI(0, pidf[1]);
-  //   winchRight.config_kD(0, pidf[2]);
-  //   winchRight.config_kF(0, pidf[3]);
-  // }
+  public void setPidsRight(double[] pidf) {
+    winchRight.config_kP(0, pidf[0]);
+    winchRight.config_kI(0, pidf[1]);
+    winchRight.config_kD(0, pidf[2]);
+    winchRight.config_kF(0, pidf[3]);
+  }
 
   // sets the encoder to 0
-  // public void zeroRight() {
-  //   winchRight.setSelectedSensorPosition(0);
-  // }
+  public void zeroRight() {
+    winchRight.setSelectedSensorPosition(0);
+  }
 
   // sets the encoder to 0
   public void zeroLeft() {
@@ -112,19 +104,17 @@ public class ClimberL extends SubsystemBase {
     return winchLeft.getSupplyCurrent();
   }
 
+  public double getRightDistance() {
+    return winchRight.getSelectedSensorPosition();
+  }
 
-
-  // public double getRightDistance() {
-  //   return winchRight.getSelectedSensorPosition();
-  // }
-
-  // public double getRightSpeed() {
-  //   return winchRight.getSelectedSensorVelocity();
-  // }
+  public double getRightSpeed() {
+    return winchRight.getSelectedSensorVelocity();
+  }
   
-  // public double getRightCurrent() {
-  //   return winchRight.getSupplyCurrent();
-  // }
+  public double getRightCurrent() {
+    return winchRight.getSupplyCurrent();
+  }
   
   // /**
   //  * @param power The percentage of power being sent to the left winch motor
@@ -138,9 +128,9 @@ public class ClimberL extends SubsystemBase {
   //  * @param power The percentage of power being sent to the right winch motor
   //  * @author Foz
   //  */
-  // public void setRightPower(double power) {
-  //   winchRight.set(ControlMode.PercentOutput, power);
-  // }
+  public void setRightPower(double power) {
+    winchRight.set(ControlMode.PercentOutput, power);
+  }
 
   /**
    * @param setPoint The positional setpoint for the left winch
@@ -154,21 +144,20 @@ public class ClimberL extends SubsystemBase {
    * @param setPoint the positional setpoint for the right winch
    * @author Foz
    */
-  // public void SetSetpointRight(double setPoint){
-  //   winchRight.set(ControlMode.Position, setPoint);
-  // }
+  public void SetSetpointRight(double setPoint){
+    winchRight.set(ControlMode.Position, setPoint);
+  }
 
 
   public void setMotionMagicSetpoint(double setPoint) {
     winchLeft.set(ControlMode.MotionMagic, setPoint);
-   // winchRight.set(ControlMode.MotionMagic, setPoint);
+    winchRight.set(ControlMode.MotionMagic, setPoint);
   }
 
   /**
    * Extends the large pistons
    * @author Cobob
    */
- 
   public TalonFX[] gibMotors() {
     return new TalonFX[] {winchLeft}; // winchRight};
   }
